+++ index.js
@@ -28,7 +28,6 @@ const cookieJar = new tough.CookieJar();
 const axios = require('axios') //.default;	// https://github.com/axios/axios
 axios.defaults.xsrfCookieName = undefined; // change  xsrfCookieName: 'XSRF-TOKEN' to  xsrfCookieName: undefined, we do not want this default,
 const axiosWS = axios.create({
-	// axios-cookiejar-support v2.0.2 required config
 	jar: cookieJar, //added in axios-cookiejar-support v2.0.x, see https://github.com/3846masa/axios-cookiejar-support/blob/main/MIGRATION.md
 });
 
@@ -3105,6 +3104,40 @@ class stbDevice {
 		this.televisionService.getCharacteristic(Characteristic.RemoteKey)
 			.on('set', this.setRemoteKey.bind(this));
 
+
+
+			
+		// Custom characteristics
+		// these are visible in Shortcuts with the name "Custom"
+		var hapCharacteristic = {};
+		const BASE_UUID =   "-0000-3C36-E400-3C36E4FF0012"; // a random UUID used only for my plugin's characteristics, based on 3C36E4-EOSSTB-003656123456
+		// 											 export const BASE_UUID = "-0000-1000-8000-0026BB765291"; // Apple HomeKit base UUID
+
+		// add a custom hap characteristic for the current channel id, appears as Custom in shortcuts
+		// var hapCharacteristic = new Characteristic(characteristic.displayName, characteristic.UUID, characteristic.props);
+		hapCharacteristic = new Characteristic("Current Channel Id", "00000001" + BASE_UUID, {
+			format: Characteristic.Formats.STRING,
+			perms: [Characteristic.Perms.PAIRED_READ, Characteristic.Perms.NOTIFY]
+		})
+		hapCharacteristic.value = ''; // add a default empty value 
+		hapCharacteristic.on('get', this.getCurrentChannelId.bind(this));
+		this.televisionService.addCharacteristic(hapCharacteristic); // add the Characteristic to the televisionService
+		// once added, it can be retrieved with
+		//this.televisionService.getCharacteristic('Current Channel Id')
+
+		// add a custom hap characteristic for the current channel name, appears as Custom in shortcuts
+		// var hapCharacteristic = new Characteristic(characteristic.displayName, characteristic.UUID, characteristic.props);
+		hapCharacteristic = new Characteristic("Current Channel Name", "00000002" + BASE_UUID, {
+			format: Characteristic.Formats.STRING,
+			perms: [Characteristic.Perms.PAIRED_READ, Characteristic.Perms.NOTIFY]
+		})
+		hapCharacteristic.value = ''; // add a default empty value 
+		hapCharacteristic.on('get', this.getCurrentChannelName.bind(this));
+		this.televisionService.addCharacteristic(hapCharacteristic); // add the Characteristic to the televisionService
+		// once added, it can be retrieved with
+		//this.televisionService.getCharacteristic('Current Channel Name')
+
+
 		//this.log('DEBUG:  this.televisionService')
 		//this.log(this.televisionService)
 
@@ -4368,7 +4401,7 @@ class stbDevice {
 	async setInput(input, callback) {
 		input = input ?? {} // ensure input is never null or undefined
 		if (this.config.debugLevel > 1) { this.log.warn('%s: setInput input %s %s',this.name, input.id, input.name); }
-		callback(null); // for rapid response
+		callback(); // for rapid response
 		var currentChannelName = NO_CHANNEL_NAME;
 		const channel = this.platform.masterChannelList.find(channel => channel.id === this.currentChannelId);
 		if (channel) { currentChannelName = channel.name; }
@@ -4393,18 +4426,13 @@ class stbDevice {
 	async getInputName(inputId, callback) {
 		// fired by the user changing a channel name in Home app accessory setup
 		//if (this.config.debugLevel > 1) { this.log.warn('%s: getInputName inputId %s', this.name, inputId); }
-
-		// need to read from stored cache, currently not implemented, TO-DO
-		var chName = NO_CHANNEL_NAME; // must have a value
-		if (this.channelList[inputId-1]) {
-			chName = this.channelList[inputId-1].configuredName;
-		}
+		const inputName =(this.channelList[inputId-1] || {}).configuredName || ''; // Empty string if not found
 		if (this.config.debugLevel > 1) { 
-			this.log.warn("%s: getInputName for input %s returning '%s'", this.name, inputId, chName);
+			this.log.warn("%s: getInputName for input %s returning '%s'", this.name, inputId, inputName);
 		}
-		callback(null, chName);
+		callback(null, inputName);
 	};
-
+	
 	// set input name (change the TV channel name)
 	async setInputName(inputId, newInputName, callback) {
 		// fired by the user changing a channel name in Home app accessory setup
@@ -4436,9 +4464,38 @@ class stbDevice {
 			//this.log('%s: Renamed channel %s from %s to %s (valid only for HomeKit)', this.name, inputId+1, oldInputName, newInputName);
 			this.log('%s: Renamed channel %s to %s (valid only for HomeKit)', this.name, inputId+1, newInputName);
 		}
-		callback(null);
+		callback();
 	};
 
+
+	// get current channel id (the TV channel identifier, a string)
+	// added in v2.1.0
+	// custom characteristic, returns a string, the event updates the characteristic value automatically
+	async getCurrentChannelId(callback, currentChannelId) {
+		// fired by the user reading the Custom characteristic in Shortcuts
+		// fired when the accessory is first created and HomeKit requests a refresh
+		// fired when the icon is clicked in the Home app and HomeKit requests a refresh
+		// fired when the Home app is opened
+		currentChannelId = this.currentChannelId || '';  // this.currentChannelId is a string eg SV09038. Empty string if not found
+		if (this.config.debugLevel > 1) { this.log.warn("%s: getCurrentChannelId returning '%s'", this.name, currentChannelId); }
+		callback(null, currentChannelId);
+	};
+
+
+	// get current channel name (the TV channel name)
+	// added in v2.1.0
+	// custom characteristic, returns a string, the event updates the characteristic value automatically
+	async getCurrentChannelName(callback, currentChannelName) {
+		// fired by the user reading the Custom characteristic in Shortcuts
+		// fired when the accessory is first created and HomeKit requests a refresh
+		// fired when the icon is clicked in the Home app and HomeKit requests a refresh
+		// fired when the Home app is opened
+		const curChannel = this.platform.masterChannelList.find(channel => channel.id === this.currentChannelId );  // this.currentChannelId is a string eg SV09038
+		// consider setting to Radio if radio is playing
+		currentChannelName = (curChannel || {}).name || ''; // Empty string if not found
+		if (this.config.debugLevel > 1) { this.log.warn("%s: getCurrentChannelName returning '%s'", this.name, currentChannelName); }
+		callback(null, currentChannelName);
+	};
 
 
 	// get input visibility state (of the TV channel in the accessory)
@@ -4469,7 +4526,7 @@ class stbDevice {
 		//not yet active
 		//this.platform.persistConfig(this.deviceId, this.channelList);
 
-		callback(null); // for rapid response
+		callback(); // for rapid response
 	}
 
 
@@ -4490,7 +4547,7 @@ class stbDevice {
 		if(this.currentClosedCaptionsState !== targetClosedCaptionsState){
 			this.log("setClosedCaptions: not yet implemented");
 		}
-		callback(null);
+		callback();
 	}
 
 
@@ -4521,7 +4578,7 @@ class stbDevice {
 		if(this.customPictureMode !== targetPictureMode){
 			this.log("setPictureMode: not yet implemented");
 		}
-		callback(null);
+		callback();
 	}